CSS
Articulation order taken directly from the CSS manual. Standard legato not included (sorry, but the advanced legato is _that_ good). Also the Con Sordino cannot be a separate articulation, since it is basically something that can be present for all other articulations, so instead of duplicating all articulations and adding '+ CS' to the articulation name, it can still be operated normally for instance via CC 58 (or any other means the library provides, like a keyswitch). Although, maybe I'll add that later if it proves to be handy (?). But for now it feels more intuitive to have it as a separate thing that can be easily switched on/off instead of fiddling around with the articulations. This is all about making things easy, you know, so why make them hard.